Add files via upload
Write a program that prompts the user for the radius of a sphere (as a float)
and displays the diameter, surface area, and volume of the sphere. All solutions
should be printed out to the screen using two decimal places. The program
itself should prompt the user for the input.