Add files via upload
Write a program that accepts an integer N from the user as a command line
argument (the program should not prompt for the input as in the previous
problem) and outputs an ASCII art image made up of N lines to the screen.
The image should be a triangle made up of * characters, with the center axis
being made up of + characters. 

Q3
Write a program that displays the first k elements of the Fibonacci sequence to
the terminal

Write a C function that takes two integer arrays (A and B) and sums them
together (element wise). A third array to accept the result/output should be
passed in as the O argument. Assume the arrays are all the same size. The
argument N is the size of the arrays.

Write a program that acts as an integer calculator that supports a single arithmetic operation at a time. The user should enter the calculation on a single
line. No values should be passed as command-line arguments.